# api/openapi.yaml
components:
  schemas:
...
    AlertRuleBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/AlertRuleBody.json
          format: uri
          readOnly: true
          type: string
        baseline:
          type: string
        enabled:
          type: boolean
        name:
          type: string
        repository:
          type: string
        run_reason:
          type: string
        threshold:
          format: double
          type: number
        threshold_z:
          format: double
          type: number
      required:
        - name
        - repository
      type: object
    AlertRuleView:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/AlertRuleView.json
          format: uri
          readOnly: true
          type: string
        baseline:
          type: string
        created_at:
          format: date-time
          type: string
        enabled:
          type: boolean
        id:
          type: string
        last_evaluated_at:
          format: date-time
          type: string
        name:
          type: string
        repository:
          type: string
        run_reason:
          type: string
        state:
          type: string
        threshold:
          format: double
          type: number
        threshold_z:
          format: double
          type: number
        updated_at:
          format: date-time
          type: string
        user_id:
          type: string
      required:
        - id
        - user_id
        - name
        - repository
        - baseline
        - threshold
        - threshold_z
        - enabled
        - state
        - created_at
        - updated_at
      type: object
...
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    ListAlertEventsOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ListAlertEventsOutputBody.json
          format: uri
          readOnly: true
          type: string
        events:
          items:
            $ref: "#/components/schemas/AlertEventView"
          type:
            - array
            - "null"
      required:
        - events
      type: object
    ListAlertRulesOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ListAlertRulesOutputBody.json
          format: uri
          readOnly: true
          type: string
        rules:
          items:
            $ref: "#/components/schemas/AlertRuleView"
          type:
            - array
            - "null"
      required:
        - rules
      type: object
...
    ResultDetail:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ResultDetail.json
          format: uri
          readOnly: true
          type: string
        batch_id:
          type:
            - string
            - "null"
        change_annotations:
          additionalProperties: {}
          type: object
        commit:
          $ref: "#/components/schemas/Commit"
        commit_repo_url:
          type: string
        context:
          additionalProperties: {}
          type: object
        data:
          items:
            format: double
            type:
              - number
              - "null"
          type:
            - array
            - "null"
        error:
          additionalProperties: {}
          type:
            - object
            - "null"
        hardware:
          $ref: "#/components/schemas/Hardware"
        history_fingerprint:
          type: string
        id:
          type: string
        info:
          additionalProperties: {}
          type: object
        iterations:
          format: int32
          type:
            - integer
            - "null"
        less_is_better:
          type:
            - boolean
            - "null"
        optional_benchmark_info:
          additionalProperties: {}
          type:
            - object
            - "null"
        run_id:
          type: string
        run_reason:
          type:
            - string
            - "null"
        run_tags:
          additionalProperties: {}
          type: object
        single_value_summary:
          format: double
          type:
            - number
            - "null"
        single_value_summary_type:
          type: string
        stats:
          $ref: "#/components/schemas/Aggregates"
        tags:
          additionalProperties: {}
          type: object
        time_unit:
          type:
            - string
            - "null"
        times:
          items:
            format: double
            type:
              - number
              - "null"
          type:
            - array
            - "null"
        timestamp:
          format: date-time
          type: string
        unit:
          type:
            - string
            - "null"
        validation:
          additionalProperties: {}
          type:
            - object
            - "null"
      required:
        - id
        - run_id
        - run_tags
        - run_reason
        - batch_id
        - timestamp
        - commit_repo_url
        - history_fingerprint
        - tags
        - context
        - info
        - hardware
        - commit
        - unit
        - less_is_better
        - time_unit
        - iterations
        - data
        - times
        - stats
        - error
        - optional_benchmark_info
        - validation
        - change_annotations
        - single_value_summary
        - single_value_summary_type
      type: object
...
    ResultPage:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ResultPage.json
          format: uri
          readOnly: true
          type: string
        next_page_cursor:
          type:
            - string
            - "null"
        results:
          items:
            $ref: "#/components/schemas/ResultListItem"
          type:
            - array
            - "null"
      required:
        - results
        - next_page_cursor
      type: object
...
info:
  title: Conbench
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/alert-rules:
    get:
      operationId: list-alert-rules
      parameters:
        - description: Bearer token.
          in: header
          name: Authorization
          schema:
            description: Bearer token.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAlertRulesOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List the caller's alert rules
    post:
      operationId: create-alert-rule
      parameters:
        - description: Bearer token.
          in: header
          name: Authorization
          schema:
            description: Bearer token.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertRuleBody"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRuleView"
          description: Created
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Create an alert rule
  /api/alert-rules/{id}:
    delete:
      operationId: delete-alert-rule
      parameters:
        - description: Bearer token.
          in: header
          name: Authorization
          schema:
            description: Bearer token.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Delete one alert rule
    get:
      operationId: get-alert-rule
      parameters:
        - description: Bearer token.
          in: header
          name: Authorization
          schema:
            description: Bearer token.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRuleView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get one alert rule
    put:
      operationId: update-alert-rule
      parameters:
        - description: Bearer token.
          in: header
          name: Authorization
          schema:
            description: Bearer token.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertRuleBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRuleView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Update one alert rule
  /api/alert-rules/{id}/events:
    get:
      operationId: list-alert-events
      parameters:
        - description: Bearer token.
          in: header
          name: Authorization
          schema:
            description: Bearer token.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
        - description: Maximum events to return.
          explode: false
          in: query
          name: limit
          schema:
            default: 50
            description: Maximum events to return.
            format: int32
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAlertEventsOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List alert events for one rule
...
  /api/benchmark-results:
    get:
      operationId: list-benchmark-results
      parameters:
        - description: Filter by run id.
          explode: false
          in: query
          name: run_id
          schema:
            description: Filter by run id.
            type: string
        - description: Filter by batch id.
          explode: false
          in: query
          name: batch_id
          schema:
            description: Filter by batch id.
            type: string
        - description: Filter by run reason.
          explode: false
          in: query
          name: run_reason
          schema:
            description: Filter by run reason.
            type: string
        - description: Lower bound (inclusive), RFC3339.
          explode: false
          in: query
          name: earliest_timestamp
          schema:
            description: Lower bound (inclusive), RFC3339.
            type: string
        - description: Upper bound (inclusive), RFC3339.
          explode: false
          in: query
          name: latest_timestamp
          schema:
            description: Upper bound (inclusive), RFC3339.
            type: string
        - description: Pagination cursor (previous page's last id).
          explode: false
          in: query
          name: cursor
          schema:
            description: Pagination cursor (previous page's last id).
            type: string
        - description: Page size (max 1000).
          explode: false
          in: query
          name: page_size
          schema:
            default: 100
            description: Page size (max 1000).
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultPage"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List and search benchmark results
  /api/benchmark-results/{id}:
    delete:
      operationId: delete-result
      parameters:
        - description: Bearer token, e.g. 'Bearer <token>'.
          in: header
          name: Authorization
          schema:
            description: Bearer token, e.g. 'Bearer <token>'.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Delete a benchmark result
    get:
      operationId: get-benchmark-result
      parameters:
        - description: Benchmark result id.
          in: path
          name: id
          required: true
          schema:
            description: Benchmark result id.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultDetail"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get a benchmark result
    put:
      operationId: update-result
      parameters:
        - description: Bearer token, e.g. 'Bearer <token>'.
          in: header
          name: Authorization
          schema:
            description: Bearer token, e.g. 'Bearer <token>'.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResultInputBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultDetail"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Update a benchmark result's change_annotations
  /api/ci/report:
    get:
      operationId: get-ci-report
      parameters:
        - description: Repository URL.
          explode: false
          in: query
          name: repository
          schema:
            description: Repository URL.
            type: string
        - description: Commit SHA.
          explode: false
          in: query
          name: commit_sha
          schema:
            description: Commit SHA.
            type: string
        - description: Comma-separated contender run IDs.
          explode: false
          in: query
          name: run_ids
          schema:
            description: Comma-separated contender run IDs.
            type: string
        - description: Comma-separated explicit baseline run IDs, paired by position with run_ids.
          explode: false
          in: query
          name: baseline_run_ids
          schema:
            description: Comma-separated explicit baseline run IDs, paired by position with run_ids.
            type: string
        - description: Automatic baseline selector.
          explode: false
          in: query
          name: baseline
          schema:
            description: Automatic baseline selector.
            enum:
              - fork_point
              - parent
              - latest_default
            type: string
        - description: Pairwise percent-change threshold. Defaults to 5.
          explode: false
          in: query
          name: threshold
          schema:
            description: Pairwise percent-change threshold. Defaults to 5.
            format: double
            type: number
        - description: Lookback z-score threshold. Defaults to 5.
          explode: false
          in: query
          name: threshold_z
          schema:
            description: Lookback z-score threshold. Defaults to 5.
            format: double
            type: number
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CIReport"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get a PR/CI benchmark report
  /api/compare/benchmark-results:
    get:
      operationId: compare-benchmark-results
      parameters:
        - description: Baseline benchmark result id.
          explode: false
          in: query
          name: baseline_result_id
          required: true
          schema:
            description: Baseline benchmark result id.
            type: string
        - description: Contender benchmark result id.
          explode: false
          in: query
          name: contender_result_id
          required: true
          schema:
            description: Contender benchmark result id.
            type: string
        - description: Pairwise percent-change threshold.
          explode: false
          in: query
          name: threshold
          schema:
            default: 5
            description: Pairwise percent-change threshold.
            format: double
            type: number
        - description: Lookback z-score threshold.
          explode: false
          in: query
          name: threshold_z
          schema:
            default: 5
            description: Lookback z-score threshold.
            format: double
            type: number
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompareResult"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Compare two benchmark results
  /api/history:
    get:
      operationId: get-history
      parameters:
        - description: History fingerprint.
          explode: false
          in: query
          name: fingerprint
          required: true
          schema:
            description: History fingerprint.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistorySeries"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get a history series by fingerprint
  /api/history/{benchmark_result_id}:
    get:
      operationId: get-history-for-result
      parameters:
        - description: Benchmark result id whose history to load.
          in: path
          name: benchmark_result_id
          required: true
          schema:
            description: Benchmark result id whose history to load.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistorySeries"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Get the history series for a benchmark result
...
  /api/results:
    post:
      operationId: submit-result
      parameters:
        - description: Bearer token, e.g. 'Bearer <token>'.
          in: header
          name: Authorization
          schema:
            description: Bearer token, e.g. 'Bearer <token>'.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitOutputBody"
          description: Created
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Submit a benchmark result
...
  /api/series:
    get:
      operationId: list-series
      parameters:
        - description: Substring match on case name and case tags.
          explode: false
          in: query
          name: q
          schema:
            description: Substring match on case name and case tags.
            type: string
        - description: Filter by hardware name.
          explode: false
          in: query
          name: hardware
          schema:
            description: Filter by hardware name.
            type: string
        - description: Filter by repository URL.
          explode: false
          in: query
          name: repository
          schema:
            description: Filter by repository URL.
            type: string
        - description: Filter by history fingerprint.
          explode: false
          in: query
          name: fingerprint
          schema:
            description: Filter by history fingerprint.
            type: string
        - description: Latest commit at or after this instant, RFC3339.
          explode: false
          in: query
          name: active_since
          schema:
            description: Latest commit at or after this instant, RFC3339.
            type: string
        - description: Latest commit at or before this instant, RFC3339.
          explode: false
          in: query
          name: active_until
          schema:
            description: Latest commit at or before this instant, RFC3339.
            type: string
        - description: Pagination cursor from a previous page's next_page_cursor.
          explode: false
          in: query
          name: cursor
          schema:
            description: Pagination cursor from a previous page's next_page_cursor.
            type: string
        - description: Page size (max 1000).
          explode: false
          in: query
          name: page_size
          schema:
            default: 100
            description: Page size (max 1000).
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesPage"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List and search benchmark series
  /api/tokens:
    get:
      operationId: list-tokens
      parameters:
        - description: Bearer token.
          in: header
          name: Authorization
          schema:
            description: Bearer token.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTokensOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: List the caller's API tokens
    post:
      operationId: create-token
      parameters:
        - description: Bearer token.
          in: header
          name: Authorization
          schema:
            description: Bearer token.
            type: string
        - in: cookie
          name: conbench_session
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenInputBody"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTokenOutputBody"
          description: Created
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Mint an API token
...
  /api/users/me:
    get:
      operationId: users-me
      parameters:
        - in: cookie
          name: conbench_session
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Current session identity